refactor: Integer 관련 메소드 IntegerUtil class로 분리
프로젝트 전역에서 사용되는 util 패키지 생성
Integer인지 판단하는 메소드, char을 Integer로 변환하는 메소드를 이곳으로 옮김
